Add permissions to breakdown templates

+++ app/Http/Controllers/BreakdownTemplateController.php
@@ -16,6 +16,11 @@ class BreakdownTemplateController extends Controller
 
     public function index()
     {
+        if (\Gate::denies('read', 'breakdown-template')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $filter = new BreakdownTemplateFilter(BreakdownTemplate::query(), session('filters.breakdown-template'));
         $breakdownTemplates = $filter->filter()->paginate(50);
         return view('breakdown-template.index', compact('breakdownTemplates'));
@@ -23,12 +28,21 @@ class BreakdownTemplateController extends Controller
 
     public function create()
     {
+        if (\Gate::denies('write', 'breakdown-template')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $wbsTree = WbsLevel::tree()->get();
         return view('breakdown-template.create')->with(['wbsTree' => $wbsTree]);
     }
 
     public function store(Request $request)
     {
+        if (\Gate::denies('write', 'breakdown-template')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
 
         if ($request->project_id) {
             $parent = BreakdownTemplate::find($request->parent_template_id);
@@ -55,16 +69,31 @@ class BreakdownTemplateController extends Controller
 
     public function show(BreakdownTemplate $breakdown_template)
     {
+        if (\Gate::denies('read', 'breakdown-template')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('breakdown-template.show', compact('breakdown_template'));
     }
 
     public function edit(BreakdownTemplate $breakdown_template)
     {
+        if (\Gate::denies('write', 'breakdown-template')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('breakdown-template.edit', compact('breakdown_template'));
     }
 
     public function update(BreakdownTemplate $breakdown_template, Request $request)
     {
+        if (\Gate::denies('write', 'breakdown-template')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, $this->rules);
 
         $breakdown_template->update($request->all());
@@ -76,6 +105,11 @@ class BreakdownTemplateController extends Controller
 
     public function destroy(BreakdownTemplate $breakdown_template)
     {
+        if (\Gate::denies('delete', 'breakdown-template')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $breakdown_template->resources()->delete();
         $breakdown_template->delete();
         flash('Breakdown template has been deleted', 'success');
@@ -88,6 +122,11 @@ class BreakdownTemplateController extends Controller
 
     function filters(Request $request)
     {
+        if (\Gate::denies('read', 'breakdown-template')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $data = $request->only(['name', 'resource_id']);
         \Session::set('filters.breakdown-template', $data);
         return \Redirect::back();
@@ -95,11 +134,21 @@ class BreakdownTemplateController extends Controller
 
     function import()
     {
+        if (\Gate::denies('write', 'breakdown-template')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('breakdown-template.import');
     }
 
     function postImport(Request $request)
     {
+        if (\Gate::denies('write', 'breakdown-template')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, [
             'file' => 'required|file|mimes:xls,xlsx',
         ]);
